+++ public/locales/en/signup.json
@@ -37,6 +37,8 @@
   "next-step": "Next step",
   "create-account": "Create account",
   "payment-info": "Payment information",
+  "payment-methods": "Payment methods",
+  "click-here": "Click here to open",
   "select-payment-plan": "Select your payment plan:",
   "what-you-will-get": "What you will get:",
   "contact-information": "Contact information",
