+++ public/locales/es/signup.json
@@ -36,6 +36,8 @@
   "next-step": "Siguiente paso",
   "create-account": "Crear cuenta",
   "payment-info": "Información de pago",
+  "payment-methods": "Métodos de pago",
+  "click-here": "Click aquí para abrir",
   "select-payment-plan": "Selecciona un plan de pago:",
   "what-you-will-get": "Lo que obtendrás:",
   "free-trial": "Prueba gratuita",
