+++ src/common/store/types/index.js
@@ -18,6 +18,9 @@ const IS_LOADING = 'IS_LOADING';
 const SET_SERVICE_PROPS = 'SET_SERVICE_PROPS';
 const SET_SELECTED_SERVICE = 'SET_SELECTED_SERVICE';
 const SET_PAYMENT_METHODS = 'SET_PAYMENT_METHODS';
+const SET_PAYMENT_STATUS = 'SET_PAYMENT_STATUS';
+const SET_SUBMITTING_CARD = 'SET_SUBMITTING_CARD';
+const SET_SUBMITTING_PAYMENT = 'SET_SUBMITTING_PAYMENT';
 const SET_COHORT_SESSION = 'SET_COHORT_SESSION';
 const SET_SORTED_ASSIGNMENTS = 'SET_SORTED_ASSIGNMENTS';
 const SET_TASK_COHORT_NULL = 'SET_TASK_COHORT_NULL';
@@ -34,6 +37,9 @@ export {
   SET_CHECKOUT_DATA,
   SET_LOCATION,
   SET_PAYMENT_INFO,
+  SET_PAYMENT_STATUS,
+  SET_SUBMITTING_CARD,
+  SET_SUBMITTING_PAYMENT,
   HANDLE_PAYMENT,
   SET_PLAN_DATA,
   SET_LOADER,
