+++ src/js_modules/checkout/ServiceSummary.jsx
@@ -3,19 +3,17 @@ import { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { Avatar, Box, Button, Link } from '@chakra-ui/react';
 import useTranslation from 'next-translate/useTranslation';
-import * as Yup from 'yup';
-import { Form, Formik } from 'formik';
 import bc from '../../common/services/breathecode';
 import useSignup from '../../common/store/actions/signupAction';
 import Heading from '../../common/components/Heading';
 import Icon from '../../common/components/Icon';
 import useStyle from '../../common/hooks/useStyle';
 import Text from '../../common/components/Text';
-import FieldForm from '../../common/components/Forms/FieldForm';
 import { formatPrice, getStorageItem } from '../../utils';
 import ModalInfo from '../moduleMap/modalInfo';
 import useCohortHandler from '../../common/hooks/useCohortHandler';
 import modifyEnv from '../../../modifyEnv';
+import CardForm from './CardForm';
 import ModalCardError from './ModalCardError';
 import { SILENT_CODE } from '../../lib/types';
 import { reportDatalayer } from '../../utils/requests';
@@ -24,45 +22,21 @@ function ServiceSummary({ service }) {
   const BREATHECODE_HOST = modifyEnv({ queryString: 'host', env: process.env.BREATHECODE_HOST });
   const { t } = useTranslation('signup');
   const {
-    state, setSelectedService, setPaymentInfo,
+    state, setSelectedService, setIsSubmittingCard, setIsSubmittingPayment,
   } = useSignup();
-  const { selectedService, paymentInfo } = state;
+  const { selectedService, isSubmittingCard } = state;
   const [confirmationOpen, setConfirmationOpen] = useState(false);
   const [purchaseCompleted, setPurchaseCompleted] = useState(false);
   const { state: cohortState } = useCohortHandler();
   const { cohortSession } = cohortState;
   const { backgroundColor, lightColor, hexColor, backgroundColor3 } = useStyle();
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [isSubmittingCard, setIsSubmittingCard] = useState(false);
   const [openDeclinedModal, setOpenDeclinedModal] = useState(false);
   const [declinedModalProps, setDeclinedModalProps] = useState({
     title: '',
     description: '',
   });
-  const [stateCard, setStateCard] = useState({
-    card_number: 0,
-    exp_month: 0,
-    exp_year: 0,
-    cvc: 0,
-  });
-  const redirectedFrom = getStorageItem('redirected-from');
 
-  const infoValidation = Yup.object().shape({
-    owner_name: Yup.string()
-      .min(6, t('validators.owner_name-min'))
-      .required(t('validators.owner_name-required')),
-    card_number: Yup.string()
-      .min(16)
-      .max(20)
-      .required(t('validators.card_number-required')),
-    exp: Yup.string()
-      .min(5, t('validators.exp-min'))
-      .required(t('validators.exp-required')),
-    cvc: Yup.string()
-      .min(3)
-      .max(4)
-      .required(t('validators.cvc-required')),
-  });
+  const redirectedFrom = getStorageItem('redirected-from');
 
   const dataToAssign = {
     service: service?.service?.slug,
@@ -103,7 +77,7 @@ function ServiceSummary({ service }) {
 
   const handlePaymentErrors = (data, actions = {}, callback = () => {}) => {
     const silentCode = data?.silent_code;
-    setIsSubmitting(false);
+    setIsSubmittingPayment(false);
     actions?.setSubmitting(false);
     callback();
     if (silentCode === SILENT_CODE.CARD_ERROR) {
@@ -455,15 +429,15 @@ function ServiceSummary({ service }) {
                   {t('select-payment-plan')}
                 </Heading>
                 <Box display="flex" flexDirection="column" gridGap="10px">
-                  <Formik
-                    initialValues={{
-                      owner_name: '',
-                      card_number: '',
-                      exp: '',
-                      cvc: '',
+                  <CardForm
+                    modalCardErrorProps={{
+                      disableTryAgain: true,
+                      openDeclinedModal,
+                      setOpenDeclinedModal,
+                      declinedModalProps,
                     }}
-                    onSubmit={(values, actions) => {
-                      setIsSubmitting(true);
+                    onSubmit={(values, actions, stateCard) => {
+                      setIsSubmittingPayment(true);
                       setIsSubmittingCard(true);
                       const monthAndYear = values.exp?.split('/');
                       const expMonth = monthAndYear[0];
@@ -477,102 +451,7 @@ function ServiceSummary({ service }) {
                       };
                       handleSubmit(actions, allValues);
                     }}
-                    validationSchema={infoValidation}
-                  >
-                    {() => (
-                      <Form
-                        style={{
-                          display: 'flex',
-                          flexDirection: 'column',
-                          gridGap: '22px',
-                        }}
-                      >
-                        <Box display="flex" gridGap="18px">
-                          <FieldForm
-                            type="text"
-                            name="owner_name"
-                            externValue={paymentInfo.owner_name}
-                            handleOnChange={(e) => {
-                              setPaymentInfo('owner_name', e.target.value);
-                              setStateCard({ ...stateCard, owner_name: e.target.value });
-                            }}
-                            pattern="[A-Za-z ]*"
-                            label={t('owner-name')}
-                          />
-                        </Box>
-                        <Box display="flex" gridGap="18px">
-                          <FieldForm
-                            type="text"
-                            name="card_number"
-                            externValue={paymentInfo.card_number}
-                            handleOnChange={(e) => {
-                              const value = e.target.value.replace(/\s/g, '').replace(/[^0-9]/g, '');
-                              const newValue = value.replace(/(.{4})/g, '$1 ').trim();
-                              e.target.value = newValue.slice(0, 19);
-                              setPaymentInfo('card_number', e.target.value);
-                              setStateCard({ ...stateCard, card_number: newValue.replaceAll(' ', '').slice(0, 16) });
-                            }}
-                            pattern="[0-9]*"
-                            label={t('card-number')}
-                          />
-                        </Box>
-                        <Box display="flex" gridGap="18px">
-                          <Box display="flex" gridGap="18px" flex={1}>
-                            <Box display="flex" flexDirection="column" flex={0.5}>
-                              <FieldForm
-                                style={{ flex: 0.5 }}
-                                type="text"
-                                name="exp"
-                                externValue={paymentInfo.exp}
-                                maxLength={3}
-                                handleOnChange={(e) => {
-                                  let { value } = e.target;
-                                  if ((value.length === 2 && paymentInfo.exp?.length === 1)) {
-                                    value += '/';
-                                  } else if (value.length === 2 && paymentInfo.exp?.length === 3) {
-                                    value = value.slice(0, 1);
-                                  }
-                                  value = value.substring(0, 5);
-
-                                  setPaymentInfo('exp', value);
-                                }}
-                                pattern="[0-9]*"
-                                label={t('exp')}
-                              />
-                            </Box>
-                            <FieldForm
-                              style={{ flex: 0.5 }}
-                              type="text"
-                              name="cvc"
-                              externValue={paymentInfo.cvc}
-                              maxLength={3}
-                              handleOnChange={(e) => {
-                                const value = e.target.value.replace(/\s/g, '').replace(/[^0-9]/g, '');
-                                const newValue = value.replace(/(.{4})/g, '$1 ').trim();
-                                e.target.value = newValue.slice(0, 4);
-
-                                setPaymentInfo('cvc', e.target.value);
-                              }}
-                              pattern="[0-9]*"
-                              label={t('cvc')}
-                            />
-                          </Box>
-                        </Box>
-
-                        <Button
-                          type="submit"
-                          width="100%"
-                          variant="default"
-                          isDisabled={!selectedService?.id}
-                          isLoading={isSubmitting}
-                          height="40px"
-                          mt="0"
-                        >
-                          {t('common:proceed-to-payment')}
-                        </Button>
-                      </Form>
-                    )}
-                  </Formik>
+                  />
                 </Box>
               </Box>
             </Box>
@@ -617,7 +496,7 @@ function ServiceSummary({ service }) {
               handlerText={t('common:confirm')}
               actionHandler={() => handlePayConsumable()}
               onClose={() => {
-                setIsSubmitting(false);
+                setIsSubmittingPayment(false);
                 setConfirmationOpen(false);
               }}
               closeButtonStyles={{ borderRadius: '3px', color: '#0097CD', borderColor: '#0097CD' }}
